fix highlight for markdown search excerpts
