+++ packages/commonwealth/package.json
@@ -185,6 +185,7 @@
     "jsonwebtoken": "^9.0.0",
     "lodash": "^4.17.21",
     "magic-sdk": "^13.2.0",
+    "markdown-to-text": "^0.1.1",
     "marked": "^4.0.12",
     "mixpanel": "^0.14.0",
     "mixpanel-browser": "^2.45.0",
